+++ components/terminus/Blockly/BlocklyComponent.tsx
@@ -4,26 +4,11 @@ import { useEffect, useRef, createRef } from 'react';
 
 import Blockly from 'blockly/core';
 import 'blockly/blocks';
-import { javascriptGenerator } from 'blockly/javascript';
 import locale from 'blockly/msg/en';
 Blockly.setLocale(locale);
 
 import { ButtonPrimary } from '@components/ButtonPrimary';
-import { ObjectMap, CONST_OBJ_GLB_ENCR } from '@components/terminus/blocks/generator';
-
-const IGNORE_FIELDS = [CONST_OBJ_GLB_ENCR];
-
-// TODO this must come from CUE and the regexp are hacked here for generation
-const simpleRegex = '^[a-zA-Z]+$';
-const regexMap = {
-  'defs.#Letters': simpleRegex,
-  'defs.#LettersWithSpaces': simpleRegex,
-  'defs.#Alphanumerical': simpleRegex,
-  'defs.#AlphanumericalWithSpaces': simpleRegex,
-  'defs.#AlphanumericalNotAccented': simpleRegex,
-  'defs.#AlphanumericalNotAccentedWithSpaces': simpleRegex,
-  'defs.#SimpleDateFormat': '^20[0-9]{2}-[0-1][1-2]-[0-2][1-8]$', // regex restricted.
-};
+import { generateModel } from '@components/terminus/blocks/generator';
 
 function BlocklyComponent(props) {
   const blocklyDiv = useRef();
@@ -37,74 +22,6 @@ function BlocklyComponent(props) {
     return `/api/admin/terminus/models/${product}`;
   };
 
-  const generateModel = () => {
-    ObjectMap.clear();
-    // trigger the BLOCKLY processing which will run our custom code generation
-    javascriptGenerator.workspaceToCode(primaryWorkspace.current);
-    const ret = _generateCUEStructureAndJSONSchemas();
-
-    // add specific BoxyHQ imports
-    return `
-    EncryptedDefinitions: ${JSON.stringify(ret[1])}
-    ${ret[0]}
-    `;
-  };
-
-  // Rudimentary way of generating a CUE file and a JSON model for synthetic data generation
-  const _generateCUEStructureAndJSONSchemas = () => {
-    let defs = ``;
-    const encrObjects = [];
-    for (const [key, value] of Object.entries(Object.fromEntries(ObjectMap))) {
-      encrObjects.push(key as never);
-      const valuesMap = Object.fromEntries(value as any);
-
-      // DEFINITIONS
-      let definitions = ``;
-      for (const [field, values] of Object.entries(valuesMap)) {
-        if (IGNORE_FIELDS.includes(field)) {
-          continue;
-        }
-        definitions += `\n\t\t\t${field}: ${values[0]}`;
-
-        let pattern = regexMap[values[0]];
-        if (pattern == null) {
-          pattern = '.*';
-        }
-      }
-
-      // ENCRYPTION
-      let encryption = ``;
-      if (valuesMap[CONST_OBJ_GLB_ENCR] != null) {
-        encryption += `${valuesMap[CONST_OBJ_GLB_ENCR]}`;
-      } else {
-        for (const [field, values] of Object.entries(valuesMap)) {
-          encryption += `\n\t\t\t${field}: ${values[1]}`;
-        }
-        encryption = `{ ${encryption}
-        }`;
-      }
-
-      // MASKS
-      let masks = ``;
-      for (const [field, values] of Object.entries(valuesMap)) {
-        if (IGNORE_FIELDS.includes(field)) {
-          continue;
-        }
-        masks += `\n\t\t\t${field}: ${values[2]}`;
-      }
-      const objectOutput = `\n#${key}: {
-        #Definition: { ${definitions}
-        }
-        #Encryption: ${encryption}
-        #Mask_admin: { ${masks}
-        }
-      }`;
-      defs += objectOutput;
-    }
-
-    return [defs, encrObjects];
-  };
-
   const modelToXML = () => {
     const xml = Blockly.Xml.workspaceToDom(primaryWorkspace.current as any);
     const domToPretty = Blockly.Xml.domToPrettyText(xml);
@@ -113,7 +30,7 @@ function BlocklyComponent(props) {
 
   const uploadModel = async () => {
     const domToPretty = modelToXML();
-    const cueModel = generateModel();
+    const cueModel = generateModel(primaryWorkspace.current);
 
     const body = {
       cue_schema: Buffer.from(cueModel).toString('base64'),
